+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## Unreleased
+
+[195d736](195d736acfe1744669f73ed3c5c9cb7f978210e7)...[e7d7f87](e7d7f87106b13033ae4eb9d053e8ab4c25ca568d)
+
+### Changes
+
+- GH-40 Fix encoding/decoding variable length field - ([e7d7f87](e7d7f87106b13033ae4eb9d053e8ab4c25ca568d)) - GH-40 (https://github.com/eastern-oak/tjiftjaf/issues/40)
+
 ## [0.1.0] - 2025-11-02
 
 Initial release